+++ mrv2/lib/mrvWidgets/mrvLogDisplay.cpp
@@ -126,6 +126,9 @@ namespace mrv
 
     inline void LogDisplay::print(const char* x, const char style)
     {
+        if (style == 'C')
+            App::ui->uiStatusBar->copy_label(x);
+
         LogData* data = new LogData(this, x, style);
         if (main_thread != std::this_thread::get_id())
         {